requirements.txt
streamlit
scikit-learn
pandas
numpy
nltk
plotly
python-docx
PyMuPDF
